# Megatokyo.Server.UnitTest/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Megatokyo.Server",
    "description": "API for the Megatokyo application",
    "contact": {
      "name": "Edouard Biton",
      "url": "https://www.aarklendoia.com"
    },
    "version": "1.0"
  },
  "paths": {
    "/api/1.0/Chapters": {
      "get": {
        "tags": [
          "Chapters"
        ],
        "summary": "Get all chapters.",
        "operationId": "GetAllChapters",
        "responses": {
          "200": {
            "description": "Return in case the list have some chapters.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChapterOutputDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChapterOutputDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChapterOutputDTO"
                  }
                }
              }
            }
          },
          "204": {
            "description": "Return in case the list is empty."
          },
          "500": {
            "description": "Return in case of internal server error."
          }
        }
      }
    },
    "/api/1.0/Chapters/{category}": {
      "get": {
        "tags": [
          "Chapters"
        ],
        "summary": "Get a chapter by his number.",
        "operationId": "GetChapter",
        "parameters": [
          {
            "name": "category",
            "in": "path",
            "description": "Chapter's category",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return in case the chapter exists.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ChapterOutputDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChapterOutputDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChapterOutputDTO"
                }
              }
            }
          },
          "400": {
            "description": "Return in case the parameters are incorect.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Returned in case the chapter is not found.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Return in case of internal server error."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ChapterOutputDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    }
  }
}